+++ app/src/main/java/com/vultisig/wallet/ui/screens/v2/defi/tron/TronDeFiPositionsScreen.kt
@@ -1,20 +1,15 @@
 package com.vultisig.wallet.ui.screens.v2.defi.tron
 
-import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
-import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
-import androidx.compose.foundation.layout.width
 import androidx.compose.foundation.rememberScrollState
 import androidx.compose.foundation.shape.CircleShape
 import androidx.compose.foundation.shape.RoundedCornerShape
@@ -26,12 +21,8 @@ import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.draw.alpha
 import androidx.compose.ui.draw.clip
-import androidx.compose.ui.graphics.Brush
 import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.layout.ContentScale
-import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -40,8 +31,6 @@ import com.vultisig.wallet.R
 import com.vultisig.wallet.data.api.models.ResourceUsage
 import com.vultisig.wallet.data.models.VaultId
 import com.vultisig.wallet.ui.components.UiIcon
-import com.vultisig.wallet.ui.components.UiSpacer
-import com.vultisig.wallet.ui.components.library.UiPlaceholderLoader
 import com.vultisig.wallet.ui.components.v2.tab.VsTab
 import com.vultisig.wallet.ui.components.v2.tab.VsTabGroup
 import com.vultisig.wallet.ui.models.defi.TronDeFiPositionsViewModel
@@ -56,12 +45,7 @@ private const val TRON_RESOURCE_BANDWIDTH = "BANDWIDTH"
 
 private val TRON_DEFI_TABS = listOf(DeFiTab.STAKED)
 
-private val TronBannerGradientTop = Color(0x17FF060A) // rgba(255,6,10, 0.09)
-private val TronBannerGradientBottom = Color(0x00FF060A) // rgba(255,6,10, 0.00)
-private val TronBannerBorder = Color(0x2BFF060A) // rgba(255,6,10, 0.17)
 private val TronEditButtonBg = Color(0xFF061B3A)
-
-private val HIDE_BALANCE_CHARS = "• ".repeat(8).trim()
 
 @Composable
 internal fun TronDeFiPositionsScreen(
@@ -102,91 +86,16 @@ private fun TronDeFiPositionsScreenContent(
             isLoading = state.isLoading,
             totalValue = "${state.availableBalanceTrx} TRX",
             isBalanceVisible = state.isBalanceVisible,
-            selectedTab = state.selectedTab,
-            onTabSelected = onTabSelected,
-            onClickManage = onClickManage,
         )
-
-        ResourceTwoCardsRow(
-            resourceUsage =
-                ResourceUsage(
-                    availableBandwidth = state.availableBandwidth,
-                    totalBandwidth = state.totalBandwidth,
-                    availableEnergy = state.availableEnergy,
-                    totalEnergy = state.totalEnergy,
-                )
-        )
-
-        if (hasNoFrozenPositions) {
-            NoPositionsContainer()
-        }
-
-        if (state.pendingWithdrawals.isNotEmpty()) {
-            TronPendingWithdrawalsCard(withdrawals = state.pendingWithdrawals)
-        }
-    }
-}
-
-@Composable
-private fun TronDeFiBanner(
-    isLoading: Boolean,
-    totalValue: String,
-    isBalanceVisible: Boolean,
-    selectedTab: Int,
-    onTabSelected: (DeFiTab) -> Unit,
-    onClickManage: () -> Unit,
-) {
-    Column(verticalArrangement = Arrangement.spacedBy(20.dp)) {
-        Row(
-            modifier =
-                Modifier.fillMaxWidth()
-                    .height(118.dp)
-                    .clip(RoundedCornerShape(16.dp))
-                    .background(
-                        Brush.verticalGradient(
-                            colors = listOf(TronBannerGradientTop, TronBannerGradientBottom)
-                        )
-                    )
-                    .border(1.dp, TronBannerBorder, RoundedCornerShape(16.dp))
-        ) {
-            Column(
-                horizontalAlignment = Alignment.Start,
-                verticalArrangement = Arrangement.Center,
-                modifier = Modifier.fillMaxHeight().width(200.dp).padding(16.dp),
-            ) {
-                Text(
-                    text = stringResource(R.string.tron),
-                    style = Theme.brockmann.body.l.medium,
-                    color = Theme.v2.colors.text.primary,
-                )
-                UiSpacer(6.dp)
-
-                if (isLoading) {
-                    UiPlaceholderLoader(modifier = Modifier.size(width = 150.dp, height = 32.dp))
-                } else {
-                    Text(
-                        text = if (isBalanceVisible) totalValue else HIDE_BALANCE_CHARS,
-                        style = Theme.satoshi.price.title1,
-                        color = Theme.v2.colors.text.primary,
-                    )
-                }
-            }
-            Column(modifier = Modifier.fillMaxHeight().weight(1f)) {
-                Image(
-                    painter = painterResource(R.drawable.tron_banner),
-                    contentDescription = null,
-                    contentScale = ContentScale.FillHeight,
-                    modifier = Modifier.alpha(0.6f),
-                )
-            }
-        }
 
         Row(
             modifier = Modifier.fillMaxWidth(),
             horizontalArrangement = Arrangement.SpaceBetween,
             verticalAlignment = Alignment.CenterVertically,
         ) {
-            VsTabGroup(index = TRON_DEFI_TABS.indexOfFirst { it.displayNameRes == selectedTab }) {
+            VsTabGroup(
+                index = TRON_DEFI_TABS.indexOfFirst { it.displayNameRes == state.selectedTab }
+            ) {
                 TRON_DEFI_TABS.forEach { tab ->
                     tab {
                         VsTab(
@@ -207,6 +116,24 @@ private fun TronDeFiBanner(
             ) {
                 UiIcon(drawableResId = R.drawable.ic_edit_pencil, size = 16.dp)
             }
+        }
+
+        ResourceTwoCardsRow(
+            resourceUsage =
+                ResourceUsage(
+                    availableBandwidth = state.availableBandwidth,
+                    totalBandwidth = state.totalBandwidth,
+                    availableEnergy = state.availableEnergy,
+                    totalEnergy = state.totalEnergy,
+                )
+        )
+
+        if (hasNoFrozenPositions) {
+            NoPositionsContainer()
+        }
+
+        if (state.pendingWithdrawals.isNotEmpty()) {
+            TronPendingWithdrawalsCard(withdrawals = state.pendingWithdrawals)
         }
     }
 }
