+++ app/src/main/java/com/vultisig/wallet/ui/screens/v2/home/components/NotEnabledContainer.kt
@@ -6,12 +6,15 @@ import androidx.compose.foundation.layout.Column
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.vultisig.wallet.R
 import com.vultisig.wallet.ui.components.UiSpacer
@@ -28,7 +31,7 @@ internal fun NotEnabledContainer(
         Column(
             horizontalAlignment = Alignment.CenterHorizontally,
             verticalArrangement = Arrangement.Center,
-            modifier = Modifier.fillMaxWidth().padding(all = 24.dp),
+            modifier = Modifier.fillMaxWidth().padding(horizontal = 40.dp, vertical = 32.dp),
         ) {
             Image(
                 painter = painterResource(R.drawable.iconcrypto),
@@ -50,6 +53,7 @@ internal fun NotEnabledContainer(
                 style = Theme.brockmann.supplementary.footnote,
                 color = Theme.v2.colors.text.tertiary,
                 textAlign = TextAlign.Center,
+                modifier = Modifier.width(250.dp),
             )
 
             if (action != null) {
@@ -59,4 +63,13 @@ internal fun NotEnabledContainer(
             }
         }
     }
+}
+
+@Preview(showBackground = true)
+@Composable
+private fun NotEnabledContainerPreview() {
+    NotEnabledContainer(
+        title = stringResource(R.string.defi_no_positions_selected),
+        content = stringResource(R.string.defi_no_positions_selected_desc),
+    )
 }
