+++ app/src/main/res/drawable/iconcrypto.xml
@@ -1,13 +1,20 @@
 <vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="25dp"
+    android:width="24dp"
     android:height="24dp"
-    android:viewportWidth="25"
+    android:viewportWidth="24"
     android:viewportHeight="24">
   <path
-      android:pathData="M9.375,9.766L12.125,8.179C12.357,8.045 12.643,8.045 12.875,8.179L15.625,9.766C15.857,9.9 16,10.148 16,10.416V13.591C16,13.859 15.857,14.107 15.625,14.241L12.875,15.828C12.643,15.962 12.357,15.962 12.125,15.828L9.375,14.241C9.143,14.107 9,13.859 9,13.591V10.416C9,10.148 9.143,9.9 9.375,9.766Z"
-      android:fillColor="#4879FD"/>
+      android:pathData="M21.25,12C21.25,17.109 17.109,21.25 12,21.25C6.891,21.25 2.75,17.109 2.75,12C2.75,6.891 6.891,2.75 12,2.75C17.109,2.75 21.25,6.891 21.25,12Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#4879FD"
+      android:strokeLineCap="round"/>
   <path
-      android:pathData="M2.5,12C2.5,6.477 6.977,2 12.5,2C18.023,2 22.5,6.477 22.5,12C22.5,17.523 18.023,22 12.5,22C6.977,22 2.5,17.523 2.5,12ZM16.375,8.464L13.625,6.876C12.929,6.474 12.071,6.474 11.375,6.876L8.625,8.464C7.929,8.866 7.5,9.609 7.5,10.412V13.588C7.5,14.392 7.929,15.134 8.625,15.536L11.375,17.124C12.071,17.526 12.929,17.526 13.625,17.124L16.375,15.536C17.071,15.134 17.5,14.392 17.5,13.588V10.412C17.5,9.609 17.071,8.866 16.375,8.464Z"
-      android:fillColor="#4879FD"
-      android:fillType="evenOdd"/>
+      android:pathData="M11.25,7.526C11.714,7.258 12.286,7.258 12.75,7.526L15.5,9.113C15.964,9.381 16.25,9.876 16.25,10.412V13.588C16.25,14.123 15.964,14.619 15.5,14.887L12.75,16.474C12.286,16.742 11.714,16.742 11.25,16.474L8.5,14.887C8.036,14.619 7.75,14.123 7.75,13.588V10.412C7.75,9.876 8.036,9.381 8.5,9.113L11.25,7.526Z"
+      android:strokeLineJoin="round"
+      android:strokeWidth="1.5"
+      android:fillColor="#00000000"
+      android:strokeColor="#4879FD"
+      android:strokeLineCap="round"/>
 </vector>
